Revert "sudo make program multicore ;)"
This reverts commit 3fe9a50569ac51e24c9f002ee0459a0d17868d15.
Rayon does not like blocks that do disk I/O when running them
in parallel. Problem stems from the fact that invoking ffmpeg
essentially becomes an I/O task because it has to write the
results back to disk.

Better option would be to first read the files to memory and use
gstreamer API to do transcoding.